migrate DialogRef to new template
